refactor: SectionIntegrationSupporter 사용해 SectionIntegrationTest 리팩토링
